Put tfsync into kble-c2a
