+++ kble-c2a/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "kble-tfsync"
+name = "kble-c2a"
 version = "0.1.0"
 edition = "2021"
 license = "MIT"
@@ -17,3 +17,4 @@ tokio-util = { version = "0.7", features = ["codec"] }
 bytes = { version = "1" }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+clap = { version = "4", features = ["derive"] }
